+++ src/esm/index.js
@@ -2,7 +2,7 @@
 
 export { default as format } from './format/index.js'
 export { default as formatInTimeZone } from './formatInTimeZone/index.js'
+export { default as fromZonedTime } from './fromZonedTime/index.js'
 export { default as getTimezoneOffset } from './getTimezoneOffset/index.js'
 export { default as toDate } from './toDate/index.js'
-export { default as utcToZonedTime } from './utcToZonedTime/index.js'
-export { default as zonedTimeToUtc } from './zonedTimeToUtc/index.js'
+export { default as toZonedTime } from './toZonedTime/index.js'