+++ src/formatInTimeZone/index.js
@@ -1,5 +1,5 @@
 import format from '../format/index.js'
-import utcToZonedTime from '../utcToZonedTime/index.js'
+import toZonedTime from '../toZonedTime/index.js'
 import cloneDeep from 'lodash.clonedeep'
 
 /**
@@ -28,5 +28,5 @@ export default function formatInTimeZone(date, timeZone, formatStr, options) {
   var extendedOptions = cloneDeep(options || {})
   extendedOptions.timeZone = timeZone
   extendedOptions.originalDate = date
-  return format(utcToZonedTime(date, timeZone), formatStr, extendedOptions)
+  return format(toZonedTime(date, timeZone), formatStr, extendedOptions)
 }