+++ src/fp/fromZonedTime/index.d.ts
@@ -1,4 +1,4 @@
 // This file is generated automatically by `scripts/build/typings.js`. Please, don't change it.
 
-import { zonedTimeToUtcWithOptions } from 'date-fns-tz/fp'
-export = zonedTimeToUtcWithOptions
+import { fromZonedTime } from 'date-fns-tz/fp'
+export = fromZonedTime