+++ src/fp/fromZonedTime/index.js
@@ -1,8 +1,8 @@
 // This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
 
-import fn from '../../zonedTimeToUtc/index.js'
+import fn from '../../fromZonedTime/index.js'
 import { convertToFP } from '../_lib/convertToFP'
 
-var zonedTimeToUtc = convertToFP(fn, 2)
+var fromZonedTime = convertToFP(fn, 2)
 
-export default zonedTimeToUtc
+export default fromZonedTime