+++ src/fp/fromZonedTimeWithOptions/index.js
@@ -1,8 +1,8 @@
 // This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
 
-import fn from '../../utcToZonedTime/index.js'
+import fn from '../../fromZonedTime/index.js'
 import { convertToFP } from '../_lib/convertToFP'
 
-var utcToZonedTimeWithOptions = convertToFP(fn, 3)
+var fromZonedTimeWithOptions = convertToFP(fn, 3)
 
-export default utcToZonedTimeWithOptions
+export default fromZonedTimeWithOptions