+++ src/fp/toZonedTime/index.js
@@ -1,8 +1,8 @@
 // This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
 
-import fn from '../../utcToZonedTime/index.js'
+import fn from '../../toZonedTime/index.js'
 import { convertToFP } from '../_lib/convertToFP'
 
-var utcToZonedTime = convertToFP(fn, 2)
+var toZonedTime = convertToFP(fn, 2)
 
-export default utcToZonedTime
+export default toZonedTime