+++ src/fp/toZonedTimeWithOptions/index.js
@@ -1,8 +1,8 @@
 // This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
 
-import fn from '../../zonedTimeToUtc/index.js'
+import fn from '../../toZonedTime/index.js'
 import { convertToFP } from '../_lib/convertToFP'
 
-var zonedTimeToUtcWithOptions = convertToFP(fn, 3)
+var toZonedTimeWithOptions = convertToFP(fn, 3)
 
-export default zonedTimeToUtcWithOptions
+export default toZonedTimeWithOptions